% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata_aq_exclusions.R
\name{getdata_aq_exclusions}
\alias{getdata_aq_exclusions}
\title{getdata_aq_exclusions}
\usage{
getdata_aq_exclusions(arg1)
}
\arguments{
\item{arg1}{variable passed into SQL select statement to pull data and name data frame}
}
\description{
Creates unconfined / confined well evaluation data frame from
 oracle data pull. User will be prompted for the password to the FDEP
 Oracle Database GWIS.
}
\details{
Function to import well evaluation information into R data frame
}
\examples{
getdata_aq_exclusions('CA_EXCLUSIONS_2020')
   Entering 'CA_EXCLUSIONS_2020' for arg1 will produce a data frame
   containing 2020 site evaluation information for FDEP Status confined
   well selections.
}
