% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata_aq_exclusions_3yr.R
\name{getdata_aq_exclusions_3yr}
\alias{getdata_aq_exclusions_3yr}
\title{getdata_aq_exclusions_3yr}
\usage{
getdata_aq_exclusions_3yr(arg1, arg2)
}
\arguments{
\item{arg1}{Variable passed into SQL select statement to indicate which year's target population
to be used for SQL statement.}

\item{arg2}{Variable passed into SQL select statement to indicate which years' site
evaluation date to used for comparision to most recent well list frame.}
}
\description{
Creates 3 data frames of unconfined / confined well evaluation data via
 oracle data pull. User will be prompted for the password to the FDEP Oracle Database GWIS.
   1) A data frame (Exclusions) containing all well evaluations for the three year period.
   2) A data frame (well_removals) containing all wells evaluated in the three period which are no
      longer included in the traget population. SQL script compares the primary key
      for the most recent well list frame to those found in the site evaluation data
      pull for the three year period.
   3) A data frame (SiteEvaluations) containing the wells which were evaluated and present in
      the well list frame for the most recent year evaluated.
 User will be prompted for the password to the FDEP Oracle Database GWIS.
}
\details{
Function to create Exclusions, Site Evaluations and well removals data frames for 3 year combined analyses.
}
\examples{
getdata_aq_exclusions_3yr("'well_listframe_2020',"'CA18','CA19','CA20'")
   entering "'well_listframe_2020',"'CA18','CA19','CA20'" for arg1, arg2
   will produce a data frame for FDEP Status confined aquifer wells
   evaluated during 2018-2020 which no longer are in the target population.

}
