% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata_fw_exclusions.R
\name{getdata_fw_exclusions}
\alias{getdata_fw_exclusions}
\title{getdata_fw_exclusions}
\usage{
getdata_fw_exclusions(arg1)
}
\arguments{
\item{arg1}{varible passed into SQL select statement to pull data and name data frame}
}
\description{
creates flowing waters exclusions dataframe from oracle data pull
}
\details{
Function to import exclusions into R dataframe from the FDEP Oracle Database GWIS
}
\examples{

getdata_fw_exclusions('CN_EXCLUSIONS_2020')
}
