% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata_fw_results.R
\name{getdata_fw_results}
\alias{getdata_fw_results}
\title{getdata_fw_results}
\usage{
getdata_fw_results(arg2)
}
\arguments{
\item{arg2}{variable passed into SQL select statement to pull data and name data frame}
}
\description{
Creates flowing water results dataframe from oracle data pull.
User will be prompted for the password to the FDEP Oracle Database GWIS.
}
\details{
Function to import results into R dataframe from the FDEP Oracle Database GWIS
}
\examples{
getdata_fw_results("'CN18'") 
   entering "'CN18'" for arg2 will produce a dataframe for FDEP Status Canals sampled in 2018.
getdata_fw_results("'CN18','CN19','CN20'")
   entering "'CN18','CN19','CN20'" for arg2 will produce a dataframe for FDEP Status Canals sampled 2018 - 2020.

}
