% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata_fw_site_removals.R
\name{getdata_fw_site_removals}
\alias{getdata_fw_site_removals}
\title{getdata_fw_site_removals}
\usage{
getdata_fw_site_removals(
  site_evaluations = NA,
  listframe_directory = ".",
  listframe = NA,
  sites_crs = 4269
)
}
\arguments{
\item{site_evaluations}{Name of data frame with site evaluation data. Though not required,
the site evaluation data typically includes data from multiple years.
Site evaluation data should be retrieved using
FDEPgetdata::getdata_fw_exclusions}

\item{listframe_directory}{Folder path for line shapefile of listframe data to be used for analysis.
Default is ".", which indicates that the shapefile is
located in the main folder of the R project working directory.}

\item{listframe}{Name of the line shapefile of listframe data to be used for analysis.}

\item{sites_crs}{Coordinate Reference System (CRS/EPSG) Code for sites evaluation data latitude and longitude.
Default is CRS 4269 (NAD83).}
}
\description{
Identifies sites, from a user-identified data frame of site evaluation
data, which are not co-located with features in a user-supplied line GIS coverage.
Co-located is defined as points located within 50m of flowing waters line features.
There are three expected outputs.
1) Data frame in simple features (sf) object format with sites to be retained for analysis.
The data frame will be named SITES_WITHIN_50M, preceded by the two-letter abbreviation(s)
for the site evaluation data's water resource(s).
2) Data frame in sf object format with sites to be removed from the Status Network analyses.
The data frame will be named SITES_NOT_WITHIN_50M, preceded by the two-letter abbreviation(s)
for the site evaluation data's water resource(s).
3) A .CSV file saved in the R project working directory, containing sites to be removed from the
Status Network analyses. The .csv file name will be SITES_NOT_WITHIN_50M.csv, preceded by the
two-letter abbreviation(s) for the site evaluation data's water resource(s).
}
\details{
Function to identify sites which are not co-located with features in
a user-supplied line GIS coverage.
}
\examples{
getdata_fw_site_removals(site_evaluations = Exclusions,
                           listframe_directory = "./data",
                           listframe = "Cycle17_streams_coverage_2023")
		Entering information for a data frame with stream site evaluaiton data, and a stream listframe shapefile
		will produce a data frame with sites to be retained (those located within 50m of the listframe line features),
		a seconds data frame with sites to be removed (those located greater than 50m from the listframe line features),
		and a .CSV file export of the information from the data frame for the sites to be removed.

}
