% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata_lake_site_removals.R
\name{getdata_lake_site_removals}
\alias{getdata_lake_site_removals}
\title{getdata_lake_site_removals}
\usage{
getdata_lake_site_removals(
  site_evaluations = NA,
  listframe_directory = ".",
  listframe = NA,
  sites_crs = 4269
)
}
\arguments{
\item{site_evaluations}{Name of data frame with site evaluation data. The site evaluation data must
include data from multiple years (typically three).
Site evaluation data should be retrieved using
FDEPgetdata::getdata_lake_exclusions.}

\item{listframe_directory}{Folder path for polygon shapefile of listframe data to be used for analysis.
Default is ".", which indicates that the shapefile is
located in the main folder of the R project working directory.}

\item{listframe}{Name of the polygon shapefile of listframe data to be used for analysis.}

\item{sites_crs}{Coordinate Reference System (CRS/EPSG) Code for sites evaluation data latitude and longitude.
Default is CRS 4269 (NAD83).}
}
\description{
Identifies sites, from a user-identified data frame of site evaluation
data, which are not co-located with features in a user-supplied polygon GIS coverage.
Co-located is defined as site evaluation points intersecting lake polygons.
There are three expected outputs.
1) Data frame in simple features (sf) object format with sites to be retained for analysis.
For lakes resources, the data frame will be named SITES_INTERSECTS, preceded by the two-letter
abbreviation(s) for the site evaluation data's water resource(s).
2) Data frame in sf object format with sites to be removed from the Status Network analyses.
For lakes resources, the data frame will be named SITES_NOT_INTERSECTS, preceded by the
two-letter abbreviation(s) for the site evaluation data's water resource(s).
3) A .CSV file saved in the R project working directory, containing sites to be removed from the
Status Network analyses. For lakes resources, the .CSV file will be named SITES_NOT_INTERSECTS.csv,
preceded by the two-letter abbreviation(s) for the site evaluation data's water resource(s).
}
\details{
Function to identify sites which are not co-located with features in
a user-supplied polygon GIS coverage.
}
\examples{
getdata_lake_site_removals(site_evaluations = Exclusions,
                           listframe_directory = "./data",
                           listframe = "Cycle17_LargeLakes_coverage_2023")
		Entering information for a data frame with large lake site evaluation data, and a large lake listframe shapefile
		will produce a data frame with sites to be retained (those that intersect the listframe polygon features),
		a seconds data frame with sites to be removed (those that do not intersect the listframe polygon features),
		and a .CSV file export of the information from the data frame for the sites to be removed.
}
