% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata_aq_exclusions.R
\name{getdata_aq_exclusions}
\alias{getdata_aq_exclusions}
\title{getdata_aq_exclusions}
\usage{
getdata_aq_exclusions(arg1)
}
\arguments{
\item{arg1}{variable passed into SQL select statement to pull data and name data frame}
}
\description{
Creates well exclusions for unconfined / confined well dataframe
from oracle data pull. User will be prompted for the password to the FDEP Oracle Database GWIS.
}
\details{
Function to import exclusions into R dataframe from the FDEP Oracle Database GWIS
}
\examples{
getdata_aq_exclusions('CA_EXCLUSIONS_2020')
   entering 'CA_EXCLUSIONS_2020' for arg1 will produce a dataframe for FDEP Status Confined
   aquifer site exclusions for 2020 site visits.
}
