% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata_fw_exclusions.R
\name{getdata_fw_exclusions}
\alias{getdata_fw_exclusions}
\title{getdata_fw_exclusions}
\usage{
getdata_fw_exclusions(arg1)
}
\arguments{
\item{arg1}{variable passed into SQL select statement to pull data and name data frame}
}
\description{
Creates flowing waters site evaluation dataframe from oracle data pull.
User will be prompted for the password to the FDEP Oracle Database GWIS.
Total nitrogen (F.A.C. 62-302.531), total phosphorus (F.A.C. 62-302.531),
and dissolved oxygen (F.A.C. 62-302.533) criteria are added for each record
based on the corresponding nutrient watershed region and bioregion.
}
\details{
Function to import flowing waters site evaluation information into R data frame
}
\examples{
getdata_fw_exclusions('CN_EXCLUSIONS_2020')
   Entering 'CN_EXCLUSIONS_2020' for arg1 will produce a dataframe of
   2020 site evaluation information for FDEP Status Canal site selections.
}
