% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata_fw_results.R
\name{getdata_fw_results}
\alias{getdata_fw_results}
\title{getdata_fw_results}
\usage{
getdata_fw_results(arg2)
}
\arguments{
\item{arg2}{varible passed into SQL select statement to pull data and name data frame}
}
\description{
creates flowing water results dataframe from oracle data pull
}
\details{
Function to import exclusions into R dataframe from the FDEP Oracle Database GWIS
}
\examples{
getdata_fw_results("'CN2001'")
}
