% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata_lake_exclusions.R
\name{getdata_lake_exclusions}
\alias{getdata_lake_exclusions}
\title{getdata_lake_exclusions}
\usage{
getdata_lake_exclusions(arg1)
}
\arguments{
\item{arg1}{variable passed into SQL select statement to pull data and name data frame}
}
\description{
Creates lake exclusions dataframe from oracle data pull.
User will be prompted for the password to the FDEP Oracle Database GWIS.
}
\details{
Function to import exclusions into R dataframe from the FDEP Oracle Database GWIS
}
\examples{
getdata_lake_exclusions('LL_EXCLUSIONS_2020')
   entering 'LL_EXCLUSIONS_2020' for arg1 will produce a dataframe for FDEP Status large
   lake site exclusions for 2020 site visits.
}
