% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata_results.R
\name{getdata_results}
\alias{getdata_results}
\title{getdata_results}
\usage{
getdata_results(arg2)
}
\arguments{
\item{arg2}{variable passed into SQL select statement to pull data and name data frame}
}
\description{
Creates results data frame from oracle data pull.
User will be prompted for the password to the FDEP Oracle Database GWIS.
Replaces measurement values with 'NA' for those measurements having any of
 the following fatal data qualifiers: '?,O,N,T,X'.
Replaces measurement values with 'NA' for those coliform measurements which
 have reported values above the criteria (2 for fecal, 4 for total) and are
 listed as below detection ('U' qualifier).
Definitions for these codes may be found in FS 62-160.700 Table 1 (Data Qualifier Codes).
}
\details{
Function to import results into R data frame from the FDEP Oracle Database GWIS
}
\examples{
getdata_results("'CN18'")
   entering "'CN18'" for arg2 will produce a data frame for FDEP Status Canals sampled in 2018.
getdata_results("'CN18','CN19','CN20'")
   entering "'CN18','CN19','CN20'" for arg2 will produce a data frame for FDEP Status Canals sampled 2018 - 2020.

}
