% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata_trend_results.R
\name{getdata_trend_results}
\alias{getdata_trend_results}
\title{getdata_trend_results}
\usage{
getdata_trend_results(arg1, arg2, arg3)
}
\arguments{
\item{arg1}{Variable passed into SQL select statement to indicate water resources
for data pull.}

\item{arg2}{Variable passed into SQL select statement to indicate start date for data pull.}

\item{arg3}{Variable passed into SQL select statement to indicate end date for data pull.}
}
\description{
Creates dataframe of either surface or groundwater trend network site data via
 oracle data pull. User will be prompted for the password to the FDEP Oracle Database GWIS.
}
\details{
Function to pull GWIS data for trend network data analyses
Create date 02/03/2023
}
\examples{
getdata_trend_results("'AQUIFER','SPRING'","'01-OCT-1998'","'31-DEC-2022'")
   Entering "'AQUIFER','SPRING'","'01-OCT-1998'","'31-DEC-2022'" for arg1, arg2, arg3
   will produce:
   1) A data frame for all FDEP Trend Network ground water stations
   for data collect from October 1, 1998 to December 31, 2022.
   2) A dataframe of those data which are duplicated in the GWIS database. That is IF
   there is more than one value for each station, collection date, parameter combination.
   3) A dataframe of the surface water trend data which excludes the duplicates.

getdata_trend_results("'CANAL','SPRING RUN','STREAM'","'01-OCT-1998'","'31-DEC-2022'")
   Entering getdata_trend_results("'CANAL','SPRING RUN','STREAM'","'01-OCT-1998'","'31-DEC-2022'")
   will produce three dataframes, similar to the above for all FDEP Trend Network surface water stations.

getdata_trend_results("'CANAL'","'01-OCT-1998'","'31-DEC-2022'")
   Entering getdata_trend_results("'CANAL'","'01-OCT-1998'","'31-DEC-2022'") will produce three
   dataframes, similar to the above for only FDEP Trend Network canal stations.

}
